Move GrandResponsable controller
