+++ Tools/Controllers/GrandResponsableGroupeController.php
@@ -1,11 +1,13 @@
 <?php declare(strict_types = 1);
-namespace LibertAPI\Groupe\GrandResponsable;
+namespace LibertAPI\Tools\Controllers;
 
 use LibertAPI\Tools\Exceptions\MissingArgumentException;
 use LibertAPI\Tools\Interfaces;
 use Psr\Http\Message\ServerRequestInterface as IRequest;
 use Psr\Http\Message\ResponseInterface as IResponse;
-use LibertAPI\Utilisateur\UtilisateurEntite;
+use \Slim\Interfaces\RouterInterface as IRouter;
+use LibertAPI\Utilisateur;
+use LibertAPI\Groupe\GrandResponsable;
 
 /**
  * Contrôleur de grand responsable de groupes
@@ -14,23 +16,17 @@ use LibertAPI\Utilisateur\UtilisateurEntite;
  * @author Wouldsmina
  *
  * @since 1.0
- * @see \Tests\Units\GroupeController
  *
  * Ne devrait être contacté que par le routeur
  * Ne devrait contacter que le GrandResponsableRepository
  */
-final class GrandResponsableController extends \LibertAPI\Tools\Libraries\AController
+final class GrandResponsableGroupeController extends \LibertAPI\Tools\Libraries\AController
 implements Interfaces\IGetable
 {
-    /**
-     * {@inheritDoc}
-     */
-    protected function ensureAccessUser(string $order, UtilisateurEntite $utilisateur)
+    public function __construct(GrandResponsable\GrandResponsableRepository $repository, IRouter $router)
     {
-        unset($order);
-        if (!$utilisateur->isAdmin()) {
-            throw new \LibertAPI\Tools\Exceptions\MissingRightException('');
-        }
+        $this->repository = $repository;
+        $this->router = $router;
     }
 
     /**
@@ -40,14 +36,11 @@ implements Interfaces\IGetable
     {
         unset($arguments);
         try {
-            $this->ensureAccessUser(__FUNCTION__, $this->currentUser);
             $groupes = $this->repository->getList(
                 $request->getQueryParams()
             );
         } catch (\UnexpectedValueException $e) {
             return $this->getResponseNoContent($response);
-        } catch (\LibertAPI\Tools\Exceptions\MissingRightException $e) {
-            return $this->getResponseForbidden($response, $request);
         } catch (\Exception $e) {
             return $this->getResponseError($response, $e);
         }
@@ -59,11 +52,11 @@ implements Interfaces\IGetable
     /**
      * Construit le « data » du json
      *
-     * @param UtilisateurEntite $entite Responsable
+     * @param Utilisateur\UtilisateurEntite $entite Responsable
      *
      * @return array
      */
-    private function buildData(UtilisateurEntite $entite)
+    private function buildData(Utilisateur\Entite $entite)
     {
         return [
             'id' => $entite->getId(),
